Create read_speed_test.py
This is the most up to date Bokeh Version of the data display for the weather station. It uses a few different techniques from the older one. Further testing is still needed, I found that there is a supply-demand curve type time delay between bokeh plotting data points with figures versus pandas resample.